correcao de codigo e modularizacao de todos os comandos do bot
